+++ packages/schema-to-zod/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm-schema-to-zod
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.7.8
+
 ## 0.1.7
 
 ### Patch Changes